Create sessions controller to handle user's session
