+++ app/helpers/sessions_helper.rb
@@ -1,0 +1,2 @@
+module SessionsHelper
+end
